refactor(dev-server-legacy): migrate tests from mocha/chai to node:test
Replace mocha globals and chai assertions with node:test and
node:assert/strict. First TypeScript test migration in the series.

- `import { expect } from 'chai'` -> `import assert from 'node:assert/strict'`
- Add `import { describe, it } from 'node:test'`
- `expect(x).to.equal(y)` -> `assert.equal(x, y)`
- `this.timeout(10000)` -> `{ timeout: 10000 }` option on describe
- `__dirname` -> `import.meta.dirname`
- `../src/legacyPlugin.js` -> `../dist/legacyPlugin.js`
- `./userAgents.js` -> `./userAgents.ts`
- Add `--test-force-exit` to prevent test hangs from open handles
