+++ packages/dev-server-legacy/package.json
@@ -27,8 +27,8 @@
   "scripts": {
     "build": "tsc",
     "start": "wds --open --config demo/server.config.mjs",
-    "test:node": "mocha \"test/**/*.test.ts\" --require ts-node/register --reporter dot",
-    "test:watch": "mocha \"test/**/*.test.ts\" --require ts-node/register --watch --watch-files src,test"
+    "test:node": "node --test --test-force-exit test/**/*.test.ts",
+    "test:watch": "node --test --test-force-exit --watch test/**/*.test.ts"
   },
   "files": [
     "*.d.ts",
